+++ tests/testthat/test-duckR-package.R
@@ -1,3 +1,0 @@
-test_that("package can be loaded", {
-  expect_identical(utils::packageDescription("duckR")$Package, "duckR")
-})